authorize: limit access for blog post

+++ src/main/java/com/reach/blog/services/impl/BlogServiceImpl.java
@@ -5,6 +5,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.reach.blog.dto.BlogCreateDTO;
@@ -15,21 +18,28 @@ import com.reach.blog.exception.NotFoundException;
 import com.reach.blog.mapper.BlogMapper;
 import com.reach.blog.models.Blog;
 import com.reach.blog.models.Category;
+import com.reach.blog.models.User;
 import com.reach.blog.repository.BlogRepository;
 import com.reach.blog.repository.CategoryRepository;
+import com.reach.blog.repository.UserRepository;
 import com.reach.blog.services.BlogService;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class BlogServiceImpl implements BlogService {
 
     private BlogRepository blogRepository;
     private CategoryRepository categoryRepository;
+    private UserRepository userRepository;
     private BlogMapper mapper;
 
     @Autowired
-    public BlogServiceImpl(BlogRepository blogRepository, CategoryRepository categoryRepository, BlogMapper mapper) {
+    public BlogServiceImpl(BlogRepository blogRepository, CategoryRepository categoryRepository,
+            UserRepository userRepository, BlogMapper mapper) {
         this.blogRepository = blogRepository;
         this.categoryRepository = categoryRepository;
+        this.userRepository = userRepository;
         this.mapper = mapper;
     }
 
@@ -72,7 +82,6 @@ public class BlogServiceImpl implements BlogService {
     @Override
     public BlogDTO findById(Long id) {
         Blog blog = blogRepository.findById(id).orElseThrow(() -> new NotFoundException("Blog could not be found"));
-
         return mapper.mapToDTO(blog);
     }
 
@@ -93,6 +102,9 @@ public class BlogServiceImpl implements BlogService {
     @Override
     public void create(BlogCreateDTO blogDTO) {
 
+        User owner = userRepository.findByUsername(getCurrentUser())
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+
         Blog blog = new Blog();
 
         blog.setTitle(blogDTO.getTitle());
@@ -106,6 +118,7 @@ public class BlogServiceImpl implements BlogService {
                 .collect(Collectors.toList());
 
         blog.setCategories(categories);
+        blog.setUser(owner);
 
         blogRepository.save(blog);
     }
@@ -113,6 +126,7 @@ public class BlogServiceImpl implements BlogService {
     @Override
     public BlogDTO update(Long id, BlogDTO blogDTO) {
         Blog blog = blogRepository.findById(id).orElseThrow(() -> new NotFoundException("Blog could not be found"));
+        this.authorize(blog);
         blog.setTitle(blogDTO.getTitle());
         blog.setDescription(blogDTO.getDescription());
         blog.setUpdatedAt(new Date());
@@ -124,6 +138,31 @@ public class BlogServiceImpl implements BlogService {
     @Override
     public void delete(Long id) {
         Blog blog = blogRepository.findById(id).orElseThrow(() -> new NotFoundException("Blog could not be found"));
+
+        // User doesn't have permission
+        this.authorize(blog);
+        // User has permission to delete the blog post
         blogRepository.delete(blog);
+    }
+
+    /**
+     * Access denied if the user is not the owner
+     * 
+     * @param blog
+     */
+    private void authorize(Blog blog) {
+        if (!blog.getUser().getUsername().equals(this.getCurrentUser())) {
+            // User doesn't have permission
+            throw new AccessDeniedException("You do not have permission to delete this blog post");
+        }
+    }
+
+    /**
+     * Get the current authenticate user
+     */
+    private String getCurrentUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        return currentUsername;
     }
 }
